@@ -18,17 +18,17 @@
   },
   "dependencies": {
     "@contentful/rich-text-react-renderer": "^13.4.0",
-    "@contentful/rich-text-types": "^14.0.0",
+    "@contentful/rich-text-types": "^14.1.0",
     "@rebass/grid": "^6.0.0",
     "dotenv": "^8.2.0",
-    "gatsby": "^2.19.12",
+    "gatsby": "^2.19.19",
     "gatsby-image": "^2.2.40",
     "gatsby-link": "^2.2.29",
     "gatsby-plugin-layout": "^1.1.22",
-    "gatsby-plugin-offline": "^3.0.34",
+    "gatsby-plugin-offline": "^3.0.35",
     "gatsby-plugin-react-helmet": "^3.1.22",
     "gatsby-plugin-robots-txt": "^1.5.0",
-    "gatsby-plugin-rollbar": "^1.0.0",
+    "gatsby-plugin-rollbar": "^1.1.0",
     "gatsby-plugin-sharp": "^2.4.5",
     "gatsby-plugin-sitemap": "^2.2.27",
     "gatsby-plugin-styled-components": "^3.1.19",
@@ -42,7 +42,7 @@
     "querystring": "^0.2.0",
     "react": "^16.9.0",
     "react-accessible-accordion": "^3.0.1",
-    "react-dates": "16.7.0",
+    "react-dates": "21.8.0",
     "react-dom": "^16.12.0",
     "react-helmet": "^5.2.1",
     "react-intercom": "^1.0.15",
@@ -54,8 +54,8 @@
     "slick-carousel": "^1.8.1",
     "slugify": "^1.3.6",
     "string-hash": "^1.1.3",
-    "styled-components": "^4.4.1",
-    "styled-system": "^5.1.4",
+    "styled-components": "^5.0.1",
+    "styled-system": "^5.1.5",
     "tinycolor2": "^1.4.1",
     "use-onclickoutside": "^0.3.1",
     "uuid-encoder": "^1.2.0"
@@ -67,23 +67,23 @@
     "@babel/preset-react": "^7.8.3",
     "@babel/preset-typescript": "^7.8.3",
     "@babel/register": "^7.8.3",
-    "@storybook/addon-a11y": "~5.2.8",
-    "@storybook/addon-actions": "~5.2.8",
-    "@storybook/addon-info": "~5.1.11",
-    "@storybook/addon-knobs": "~5.2.8",
-    "@storybook/react": "~5.2.8",
+    "@storybook/addon-a11y": "~5.3.13",
+    "@storybook/addon-actions": "~5.3.13",
+    "@storybook/addon-info": "~5.3.13",
+    "@storybook/addon-knobs": "~5.3.13",
+    "@storybook/react": "~5.3.13",
     "@testing-library/jest-dom": "^5.0.2",
     "@testing-library/react": "^9.4.0",
     "@testing-library/react-hooks": "^3.2.1",
-    "@types/jest": "^25.1.0",
-    "@types/react": "^16.9.19",
+    "@types/jest": "^25.1.3",
+    "@types/react": "^16.9.22",
     "@types/react-dates": "^17.1.7",
     "@types/react-dom": "^16.9.5",
     "@types/storybook__addon-info": "^5.2.1",
-    "@types/styled-components": "^4.4.2",
-    "@types/styled-system": "^5.1.4",
-    "@typescript-eslint/eslint-plugin": "^2.18.0",
-    "@typescript-eslint/parser": "^2.18.0",
+    "@types/styled-components": "^5.0.0",
+    "@types/styled-system": "^5.1.7",
+    "@typescript-eslint/eslint-plugin": "^2.20.0",
+    "@typescript-eslint/parser": "^2.20.0",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
     "babel-plugin-remove-graphql-queries": "^2.7.22",
@@ -103,13 +103,13 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.18.0",
-    "eslint-plugin-react-hooks": "^2.3.0",
+    "eslint-plugin-react-hooks": "^2.4.0",
     "eslint-plugin-tsc": "^1.2.0",
     "gatsby-plugin-typescript": "^2.1.26",
     "husky": "^4.2.1",
     "inquirer": "^7.0.4",
     "jest": "^25.1.0",
-    "jest-styled-components": "^6.3.4",
+    "jest-styled-components": "^7.0.0",
     "lint-staged": "^10.0.3",
     "node-fetch": "^2.6.0",
     "prettier": "^1.19.1",
@@ -117,14 +117,14 @@
     "react-docgen-typescript-loader": "^3.6.0",
     "remark-cli": "^7.0.1",
     "remark-preset-lint-recommended": "^3.0.3",
-    "stylelint": "^13.0.0",
+    "stylelint": "^13.2.0",
     "stylelint-a11y": "^1.2.2",
     "stylelint-config-prettier": "^8.0.1",
     "stylelint-config-recommended": "^3.0.0",
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-prettier": "^1.1.2",
     "stylelint-processor-styled-components": "^1.9.0",
-    "typescript": "^3.7.5",
+    "typescript": "^3.8.2",
     "typescript-styled-plugin": "^0.15.0"
   },
   "resolutions": {
