@@ -1,6 +1,3 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import moment from 'moment'
 import {
   filterByDate,
   filterByFree,
@@ -12,6 +9,9 @@ import {
   sanitizeDates,
 } from '../../features/events/helpers/index'
 import constants from '../../constants'
+import React, { useState } from 'react'
+import PropTypes from 'prop-types'
+import moment from 'moment'
 
 const AppContext = React.createContext()
 const { Consumer } = AppContext
@@ -30,6 +30,75 @@ function getInitialFilterState() {
   }
 }
 
+const HandleEvents = events => {
+  const allEventOccurences = []
+
+  events.forEach(event => {
+    if (!event.node.recurrenceDates) {
+      allEventOccurences.push(event)
+    } else {
+      const recurrenceDates = sanitizeDates([
+        moment(event.node.startTime).format(constants.dateFormat),
+        ...event.node.recurrenceDates,
+      ])
+      const time = moment(event.node.startTime).format('HH:mm')
+      const duration = getDuration(event.node.startTime, event.node.endTime)
+
+      recurrenceDates.forEach(date => {
+        const copy = JSON.parse(JSON.stringify(event))
+
+        copy.node.startTime = moment(
+          `${date} ${time}`,
+          'DD/MM/YYYY hh:mm'
+        ).format()
+        copy.node.endTime = moment(copy.node.startTime)
+          .add(duration, 'milliseconds')
+          .format()
+
+        allEventOccurences.push(copy)
+      })
+    }
+  })
+  return allEventOccurences
+    .filter(filterPastEvents)
+    .sort((a, b) =>
+      a.node.startTime < b.node.startTime
+        ? -1
+        : a.node.startTime > b.node.startTime
+        ? 1
+        : 0
+    )
+}
+
+const filteredEvents = (events, filters) => {
+  const filteredEvents = events
+    .filter(filterByDate, {
+      startDate: filters.startDate,
+      endDate: filters.endDate,
+    })
+    .filter(filterByFree, filters.free)
+    .filter(filterByCategory, {
+      array: filters.eventCategories,
+      key: 'eventCategories',
+    })
+    .filter(filterByCategory, {
+      array: filters.venueDetails,
+      key: 'venueDetails',
+    })
+    .filter(filterByCategory, {
+      array: filters.accessibilityOptions,
+      key: 'accessibilityOptions',
+    })
+    .filter(filterByCategory, {
+      array: filters.audience,
+      key: 'audience',
+    })
+    .filter(filterByArea, filters.area)
+    .filter(filterByTime, filters.timeOfDay)
+
+  return filteredEvents
+}
+
 const initialState = {
   events: [],
   filterOpen: null,
@@ -37,60 +106,22 @@ const initialState = {
   filters: getInitialFilterState(),
 }
 
-class Provider extends Component {
-  constructor() {
-    super()
-    this.state = {
-      ...initialState,
-    }
-  }
+export const Provider = ({ events, children }) => {
+  const [state, setState] = useState({
+    ...initialState,
+    events: HandleEvents(events),
+  })
 
-  static getDerivedStateFromProps(nextProps, prevState) {
-    if (nextProps.events !== prevState.events) {
-      const allEventOccurences = []
-
-      nextProps.events.forEach(event => {
-        if (!event.node.recurrenceDates) {
-          allEventOccurences.push(event)
-        } else {
-          const recurrenceDates = sanitizeDates([
-            moment(event.node.startTime).format(constants.dateFormat),
-            ...event.node.recurrenceDates,
-          ])
-          const time = moment(event.node.startTime).format('HH:mm')
-          const duration = getDuration(event.node.startTime, event.node.endTime)
-
-          recurrenceDates.forEach(date => {
-            const copy = JSON.parse(JSON.stringify(event))
-
-            copy.node.startTime = moment(
-              `${date} ${time}`,
-              'DD/MM/YYYY hh:mm'
-            ).format()
-            copy.node.endTime = moment(copy.node.startTime)
-              .add(duration, 'milliseconds')
-              .format()
-
-            allEventOccurences.push(copy)
-          })
-        }
+  const showMore = filteredCount => {
+    if (this.state.eventsToShow < filteredCount) {
+      setState({
+        eventsToShow: state.eventsToShow + constants.itemsToLoad,
       })
-      return {
-        events: allEventOccurences
-          .filter(filterPastEvents)
-          .sort((a, b) =>
-            a.node.startTime < b.node.startTime
-              ? -1
-              : a.node.startTime > b.node.startTime
-              ? 1
-              : 0
-          ),
-      }
     }
   }
 
-  getDatepickerValues = ({ startDate, endDate }) => {
-    this.setState(prevState => ({
+  const getDatepickerValues = ({ startDate, endDate }) => {
+    setState(prevState => ({
       ...prevState,
       filters: {
         ...prevState.filters,
@@ -100,8 +131,8 @@ class Provider extends Component {
     }))
   }
 
-  setDate = (dateToSet, dateToGet) => {
-    this.setState(prevState => ({
+  const setDate = (dateToSet, dateToGet) => {
+    setState(prevState => ({
       ...prevState,
       filters: {
         ...prevState.filters,
@@ -110,8 +141,8 @@ class Provider extends Component {
     }))
   }
 
-  getCheckboxBool = (name, checked) => {
-    this.setState(prevState => ({
+  const getCheckboxBool = (name, checked) => {
+    setState(prevState => ({
       ...prevState,
       filters: {
         ...prevState.filters,
@@ -120,106 +151,69 @@ class Provider extends Component {
     }))
   }
 
-  getCheckboxSetValues = (e, name) => {
-    const state = {
-      ...this.state,
-      filters: { ...this.state.filters },
+  const getCheckboxSetValues = (e, name) => {
+    const updatedState = {
+      ...state,
+      filters: {
+        ...state.filters,
+      },
     }
 
     if (
       e.target.checked &&
-      state.filters[name].indexOf(e.target.value) === -1
+      updatedState.filters[name].indexOf(e.target.value) === -1
     ) {
-      state.filters[name].push(e.target.value)
+      updatedState.filters[name].push(e.target.value)
     } else {
-      const index = state.filters[name].indexOf(e.target.value)
+      const index = updatedState.filters[name].indexOf(e.target.value)
       if (index > -1) {
-        this.state.filters[name].splice(index, 1)
+        state.filters[name].splice(index, 1)
       }
     }
 
-    this.setState(state)
+    setState(updatedState)
   }
 
-  clearFilters = () => {
-    this.setState({
+  const clearFilters = () => {
+    setState({
       filterOpen: null,
       filters: getInitialFilterState(),
     })
   }
 
-  closeSiblingFilters = (filterName, isOpen) => {
-    if (isOpen && filterName != this.state.filterOpen) {
-      this.setState(prevState => ({
+  const closeSiblingFilters = (filterName, isOpen) => {
+    if (isOpen && filterName != state.filterOpen) {
+      setState(prevState => ({
         ...prevState,
         filterOpen: filterName,
       }))
     } else {
-      this.setState(prevState => ({
+      setState(prevState => ({
         ...prevState,
         filterOpen: null,
       }))
     }
   }
 
-  filterEvents = () => {
-    const filteredEvents = this.state.events
-      .filter(filterByDate, {
-        startDate: this.state.filters.startDate,
-        endDate: this.state.filters.endDate,
-      })
-      .filter(filterByFree, this.state.filters.free)
-      .filter(filterByCategory, {
-        array: this.state.filters.eventCategories,
-        key: 'eventCategories',
-      })
-      .filter(filterByCategory, {
-        array: this.state.filters.venueDetails,
-        key: 'venueDetails',
-      })
-      .filter(filterByCategory, {
-        array: this.state.filters.accessibilityOptions,
-        key: 'accessibilityOptions',
-      })
-      .filter(filterByCategory, {
-        array: this.state.filters.audience,
-        key: 'audience',
-      })
-      .filter(filterByArea, this.state.filters.area)
-      .filter(filterByTime, this.state.filters.timeOfDay)
-
-    return filteredEvents
-  }
-
-  showMore = filteredCount => {
-    if (this.state.eventsToShow < filteredCount) {
-      this.setState({
-        eventsToShow: this.state.eventsToShow + constants.itemsToLoad,
-      })
-    }
-  }
-
-  render() {
-    return (
-      <AppContext.Provider
-        value={{
-          state: this.state,
-          filteredEvents: this.filterEvents(),
-          actions: {
-            getCheckboxBool: this.getCheckboxBool,
-            getDatepickerValues: this.getDatepickerValues,
-            getCheckboxSetValues: this.getCheckboxSetValues,
-            clearFilters: this.clearFilters,
-            closeSiblingFilters: this.closeSiblingFilters,
-            showMore: this.showMore,
-            setDate: this.setDate,
-          },
-        }}
-      >
-        {this.props.children}
-      </AppContext.Provider>
-    )
-  }
+  return (
+    <AppContext.Provider
+      value={{
+        state,
+        filteredEvents: filteredEvents(state.events, state.filters),
+        actions: {
+          getCheckboxBool: getCheckboxBool,
+          getDatepickerValues: getDatepickerValues,
+          getCheckboxSetValues: getCheckboxSetValues,
+          clearFilters: clearFilters,
+          closeSiblingFilters: closeSiblingFilters,
+          showMore: showMore,
+          setDate: setDate,
+        },
+      }}
+    >
+      {children}
+    </AppContext.Provider>
+  )
 }
 
 Provider.propTypes = {
@@ -234,4 +228,4 @@ Provider.defaultProps = {
   events: [],
 }
 
-export { AppContext, Provider, Consumer }
+export { AppContext, HandleEvents, Consumer }
